refactor: :recycle: move bloated logic to separate hooks
